+++ cmd/helper_test.go
@@ -7,22 +7,9 @@ import (
 
 	"github.com/mschuchard/concourse-vault-resource/concourse"
 	"github.com/mschuchard/concourse-vault-resource/vault"
-	"github.com/mschuchard/concourse-vault-resource/vault/util"
 )
 
 // minimum coverage testing for helper functions
-func TestVaultClientFromSource(test *testing.T) {
-	source := concourse.Source{Token: util.VaultToken}
-	client, err := VaultClientFromSource(source)
-	if err != nil {
-		test.Error("failed to initialize vault client from concourse source")
-		test.Error(err)
-	} else if client.Token() != util.VaultToken {
-		test.Error("vault client configured with parameters from concourse source was not authenticated with expected token from source parameters")
-		test.Errorf("actual token: %s, expected token: %s", client.Token(), util.VaultToken)
-	}
-}
-
 func TestSecretsToJsonFile(test *testing.T) {
 	secretValues := concourse.SecretValues{"secretValue": {"key": "value"}}
 	if err := SecretsToJsonFile(".", secretValues); err != nil {