+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### 1.2.2 (Next)
 - Improve expiration time format in metadata version.
 - Properly utilize `aws_mount_path` `source` parameter.
+- Improve Vault client initialization efficiency.
 
 ### 1.2.1
 - Various code optimization and improvements.